refactor(silent-reply): relocate out of legacy/core to src/
silent-reply is a framework-level transport convention used by both
heartbeat and discord, not agent-core logic. Move to top-level alongside
paths.ts / config.ts.

Closes #676
