+++ src/legacy/automation/heartbeat.ts
@@ -5,7 +5,7 @@
 import fs from "node:fs/promises";
 import path from "node:path";
 import { createLogger, type Logger } from "../../logging/logger.js";
-import { isSilentReply } from "../core/silent-reply.js";
+import { isSilentReply } from "../../silent-reply.js";
 
 // ---------------------------------------------------------------------------
 // Types
